agent_bench/web_ui/backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.5
pyyaml>=6.0
sse-starlette>=1.0.0
